backend/requirements/base.txt
# Framework
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Auth
djangorestframework-simplejwt==5.3.1

# Async
celery==5.3.6
redis==5.0.3

# HTTP client (for OpenRouter)
httpx==0.27.0

# Utils
python-decouple==3.8
Pillow==10.3.0
